+++ snowddl/converter/database.py
@@ -22,8 +22,5 @@ class DatabaseConverter(AbstractConverter):
         object_path = self.base_path / self._normalise_name_with_prefix(row['database'])
         object_path.mkdir(mode=0o755, parents=True, exist_ok=True)
 
-        if data:
-            self._dump_file(object_path / 'params.yaml', data, database_json_schema)
-            return ConvertResult.DUMP
-
-        return ConvertResult.EMPTY
+        self._dump_file(object_path / 'params.yaml', data, database_json_schema)
+        return ConvertResult.DUMP